Add scripts for test_connection_odbc on Windows
`test_connection_odbc` requires the DSN setup either in an .ini file
(on Linux and Mac) or in Windows registry.

This change adds Pyton scripts that can be used for DSN setup on
Windows the same way as existing Bash scripts are used on Linux/Mac.

Testing: no functional changes, no new tests.

Fixes: #47